Use bigint as primary key of the migration_versions table

+++ src/jennifer/adapter/base.cr
@@ -284,7 +284,7 @@ module Jennifer
         return if table_exists?(Migration::Version.table_name)
 
         tb = Migration::TableBuilder::CreateTable.new(self, Migration::Version.table_name)
-        tb.integer(:id, {:primary => true, :auto_increment => true})
+        tb.bigint(:id, {:primary => true, :auto_increment => true})
         tb.string(:version, {:size => 17, :null => false})
         tb.process
       end
